#55: Split `test_bin_xcop.rb` to multiple test files
